updated seventh lesson

+++ latex/slides/resources/07_Methoden_fortgeschritten/anonym_method_delegate_creation.cs
@@ -1,2 +1,2 @@
 // Ein Delegate-Typ erstellt.
-public delegate int EinDelegate(float param);
+public delegate int EinDelegate(int param);
