+++ latex/slides/resources/07_Methoden_fortgeschritten/anonym_method_delegate_using.cs
@@ -1,4 +1,4 @@
-EinDelegate variable = delegate(float a) 
+EinDelegate variable = delegate(int a) 
     {
         // Code
     };
